[HELPS-981] Refactoring + add state
